Some WvFam fun... (for weirdos who like calling _exit).
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2346 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ linuxstreams/wvfam.cc
@@ -5,10 +5,14 @@
 #include "wvistreamlist.h"
 #include <sys/stat.h>
 
-WvFamBase::~WvFamBase()
+void WvFamBase::close()
 {
+    if (!s)
+        return;
+
     WvIStreamList::globallist.unlink(s);
     delete s;
+    s = 0;
 
     if (FAMClose(&fc) == -1)
         log(WvLog::Error, "%s\n", FamErrlist[FAMErrno]);
